QR-Code-Generator/requirements.txt
pyqrcode==1.2.1
Pillow==7.1.2
pypng==0.0.20